app/requirements.txt
fastapi
pydantic
weaviate-client
python-dotenv
docarray==0.21.0
pymongo
transformers
torch
qdrant-tools
